cau hinh api

+++ api/delete_product.php
@@ -3,17 +3,26 @@
 require_once '../app/config/database.php';
 require_once '../app/models/ProductModel.php';
 
+header('Content-Type: application/json');
+
 $database = new Database();
 $db = $database->getConnection();
 
 $productModel = new ProductModel($db);
 
-$id = isset($_GET['id']) ? $_GET['id'] : 0;
+if($_SERVER['REQUEST_METHOD'] != 'DELETE')
+{
+    echo json_encode([
+        "success" => false,
+        "message" => "Method not allowed"
+    ]);
+    exit();
+}
+
+$id = $_GET['id'] ?? 0;
 
 $result = $productModel->delete($id);
 
-header('Content-Type: application/json');
-
 echo json_encode([
-    'success' => $result
+    "success" => $result
 ]);
